Turning lane option and simple routing algorithm
Added an option on whether vehicles can make turn in all the lanes in multi-lane roads. This new option is set to true by default. In old versions, vehicles can only turn from specific lanes, such as turning right from the right-most lane. Although this is designed to reflect realistic behaviour, it can lead to heavy congestions as OpenStreetMap data does not contain turning lanes in many areas. The new option can be configured on the GUI and the script.

The simple routing algorithm for generating random traffic is also updated for achieving more realistic routes.